pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.73.0",
    "fastapi>=0.121.3",
    "google-generativeai>=0.8.5",
    "openai>=2.8.0",
    "requests>=2.32.5",
    "typer>=0.20.0",
    "uvicorn>=0.38.0",
    "wikipedia>=1.4.0",
]
